#11 spec green | class Slide moved. emaki.rb ---> lib/slide.rb
